Move image_conversion.py into _utils/
It's an internal implementation detail of Dataset.from_images()/
add_layer_from_images(), not part of the public dataset/ namespace,
so it belongs alongside the other reader/conversion internals
(pims_images.py, chunked_images.py, ims_chunked_images.py,
mrc_chunked_images.py) in _utils/ rather than as a dataset/-level
sibling of dataset.py itself.

Updated its relative imports for the extra directory depth (one more
level up for dataset_properties/geometry/utils/defaults/layer, one
fewer for what's now sibling modules in _utils/). dataset.py's import
changed from `from . import image_conversion` to
`from ._utils import image_conversion`.
